+++ packages/microservice-utilities/kafka/KafkaController.ts
@@ -7,20 +7,39 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+import KafkaTS from './KafkaTS';
+import { KafkaMessageValue, SubscriptionCallback, Headers, KafkaTSConfig } from './types';
 
-import type { SocketMessage } from './SocketMessage';
+class KafkaController {
+  private kafka: KafkaTS;
 
-export type MessageMock = {
-  createOutgoingMessage: (incomingMessage: SocketMessage) => SocketMessage | undefined;
-  delay?: number;
-};
+  constructor(config: KafkaTSConfig) {
+    this.kafka = new KafkaTS(config);
+  }
 
-export type MessageMocks = {
-  [key: string]: MessageMock;
-};
+  public async init(): Promise<void> {
+    return this.kafka.init();
+  }
 
-export type SocketConnectorConfig = {
-  socketServerPath?: string;
-  serviceId: string;
-  responseMessageMocks?: MessageMocks;
-};
+  public async addTopicSubscription<M>(
+    topicName: string,
+    callback: SubscriptionCallback<M>
+  ): Promise<void> {
+    return this.kafka.subscribeTopic(topicName, callback);
+  }
+
+  public async run(): Promise<void> {
+    return this.kafka.runConsumer();
+  }
+
+  public async produce<M = KafkaMessageValue>(
+    topicName: string,
+    value: M,
+    headers?: Headers
+    // onError?: OnProducerError<M>
+  ): Promise<void> {
+    return this.kafka.produceMessage(topicName, value, headers);
+  }
+}
+
+export default KafkaController;