+++ packages/socket-connector/src/services/mockUtils/index.ts
@@ -7,30 +7,176 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
+import {
+  getPasswordStrength,
+  JWT_TOKEN,
+  UNVERIFIED_MAIL_JWT_TOKEN,
+  validateEmail,
+  validateToken,
+} from './authorization';
 
 import type { MessageMocks } from '../../types/socketConnector';
+import {
+  ForgotPasswordResponse,
+  LoginResponse,
+  LogoutResponse,
+  RegistrationResponse,
+  UpdatePasswordResponse,
+} from './types';
 import type { SocketMessage } from '../../types/SocketMessage';
 
 export const loginMock: MessageMocks = {
+  RegistrationRequest: {
+    createOutgoingMessage: ({ payload }: SocketMessage): RegistrationResponse => {
+      const { password } = payload;
+
+      if (getPasswordStrength(password) < 5) {
+        return {
+          type: 'RegistrationResponse',
+          payload: {
+            error: {
+              status: 42001,
+              message: 'Password does not match our criteria.',
+            },
+          },
+        };
+      }
+      return {
+        type: 'RegistrationResponse',
+        payload: {
+          data: {
+            success: true,
+          },
+        },
+      };
+    },
+    delay: 1000,
+  },
   LoginRequest: {
-    createOutgoingMessage: ({ payload }: SocketMessage): SocketMessage => {
-      const { email } = payload;
+    createOutgoingMessage: ({ payload }: SocketMessage): LoginResponse => {
+      const { email, password } = payload;
+      const isValidMail = validateEmail(email);
+      const isValidPassword = getPasswordStrength(password) < 5;
+
+      if (!isValidMail || !isValidPassword) {
+        return {
+          type: 'LoginResponse',
+          payload: {
+            error: {
+              status: 401,
+              message: 'Unauthorized!',
+            },
+          },
+        };
+      }
+
+      if (email === 'unverified@something.technology') {
+        return {
+          type: 'LoginResponse',
+          payload: {
+            data: {
+              token: UNVERIFIED_MAIL_JWT_TOKEN,
+            },
+          },
+        };
+      }
 
       return {
         type: 'LoginResponse',
         payload: {
-          success: email.includes('@ikooko.io'),
-          email,
-          expiryDate: Date.now() + 1000 * 60 * 15, // in 15mins
-          error: email.includes('@ikooko.io')
-            ? undefined
-            : {
-                status: 401,
-                message: 'Unauthorized! Please use @ikooko.io email to login.',
-              },
+          data: {
+            token: JWT_TOKEN,
+          },
         },
       };
     },
+    delay: 1000,
+  },
+  SilentLoginRequest: {
+    createOutgoingMessage: ({ payload }: SocketMessage): LoginResponse => {
+      const { token } = payload;
+      if (validateToken(token)) {
+        return {
+          type: 'LoginResponse',
+          payload: {
+            error: {
+              status: 401,
+              message: 'Unauthorized!',
+            },
+          },
+        };
+      }
+
+      return {
+        type: 'LoginResponse',
+        payload: {
+          data: {
+            token: JWT_TOKEN,
+          },
+        },
+      };
+    },
+    delay: 300,
+  },
+  LogoutRequest: {
+    createOutgoingMessage: (): LogoutResponse => ({
+      type: 'LogoutResponse',
+      payload: {
+        data: { success: true },
+      },
+    }),
+  },
+  ForgotPasswordRequest: {
+    createOutgoingMessage: ({ payload }: SocketMessage): ForgotPasswordResponse => {
+      const { email } = payload;
+
+      if (email.includes('@test.de')) {
+        return {
+          type: 'ForgotPasswordResponse',
+          payload: {
+            error: {
+              status: 404,
+              message: 'No user found!',
+            },
+          },
+        };
+      }
+      return {
+        type: 'ForgotPasswordResponse',
+        payload: {
+          data: {
+            message: 'Confirmation link sent to the inbox.',
+          },
+        },
+      };
+    },
+  },
+  UpdatePasswordRequest: {
+    createOutgoingMessage: ({ payload }: SocketMessage): UpdatePasswordResponse => {
+      const { password } = payload;
+
+      if (getPasswordStrength(password) < 5) {
+        return {
+          type: 'UpdatePasswordResponse',
+          payload: {
+            error: {
+              status: 42001,
+              message: 'Password does not match our criteria.',
+            },
+          },
+        };
+      }
+
+      return {
+        type: 'UpdatePasswordResponse',
+        payload: {
+          data: {
+            message: 'Password updated successfully',
+          },
+        },
+      };
+    },
+    delay: 300,
   },
 };