+++ packages/microservice-utilities/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.14 (2021-04-28)
+
+**Note:** Version bump only for package @something.technology/microservice-utilities
+
+
+
+
+
+## 0.0.13 (2021-04-27)
+
+**Note:** Version bump only for package @something.technology/microservice-utilities
+
+
+
+
+
+## 0.0.12 (2021-04-15)
+
+**Note:** Version bump only for package @something.technology/microservice-utilities
+
+
+
+
+
+## 0.0.11 (2021-04-14)
+
+
+### Bug Fixes
+
+* add redux-connector and socket-connector libraries to readme ([c50f8c4](https://github.com/Something-Technology/something-ts/commit/c50f8c43d710ba65a2d2927ef2344cc9eddaba23))
+
+
+
+
+
+## 0.0.10 (2021-04-14)
+
+**Note:** Version bump only for package @something.technology/microservice-utilities
+
+
+
+
+
+## 0.0.9 (2021-03-23)
+
+**Note:** Version bump only for package @something.technology/microservice-utilities
+
+
+
+
+
 ## [0.0.8](https://github.com/Something-Technology/something-ts/compare/@something.technology/microservice-utilities@0.0.7...@something.technology/microservice-utilities@0.0.8) (2021-03-23)
 
 **Note:** Version bump only for package @something.technology/microservice-utilities